runtimes/pyproject.toml
[project]
name = "app"
description = "Demo of titiler.xarray."
requires-python = ">=3.8"
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]
license = {text = "MIT"}
keywords = [
    "TiTiler",
    "zarr",
    "xarray",
    "Fastapi",
]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]
dynamic = ["version"]
dependencies = [
    "titiler.xarray==0.19.0",
    "fsspec",
    "h5netcdf",
    "zarr",
    "s3fs",
    "aiohttp",
    "starlette-cramjam>=0.4,<0.5",
    "pydantic-settings~=2.0",
    "diskcache",
]

[project.optional-dependencies]
dev = [
    "pre-commit"
]
server = [
    "uvicorn"
]

[project.urls]
Homepage = "https://github.com/developmentseed/titiler-md-demo"

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
known_first_party = ["titiler"]
known_third_party = [
    "fsspec",
    "rio_tiler",
    "morecantile",
    "fastapi",
    "starlette"
]
default_section = "THIRDPARTY"

[tool.mypy]
no_strict_optional = true

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = [
    "D1",  # pydocstyle errors
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # flake8
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B905",  # ignore zip() without an explicit strict= parameter, only support with python >3.10
    "B028",
]
